### 2017-07-09 #311 @whedon
[PRE REVIEW]: Hytool: an open source matlab toolbox for the interpretation of hydraulic tests using analytical solutions
**Submitting author:** @philipperenard (<a href="http://orcid.org/0000-0003-4504-435X">Philippe Renard</a>)
**Repository:** <a href="https://github.com/UniNE-CHYN/hytool" target ="_blank">https://github.com/UniNE-CHYN/hytool</a>
**Version:** 2.05
