### 2017-07-10 #311 @kyleniemeyer
Hi @philipperenard, thanks for your submission. Do you have a suggested reviewer or two from your domain?

For existing JOSS reviewers, @nicoguaro & @pboesu, would either of you be interested to review this?
