### 2017-07-10 #311 @nicoguaro
@kyleniemeyer, I don't have access to MATLAB right now, and the repository does not specify which version is required. Can it be ran in Octave?
