### 2017-07-11 #311 @philipperenard
I do not expect that the code is fully compatible with octave. Currently it runs on R2017a and requires the Statistics toolbox.
