### 2017-07-11 #311 @nicoguaro
@kyleniemeyer, @philipperenard I don't that I have access to such a recent version of MATLAB. Not sure about the statistics toolbox, though.
