### 2017-07-12 #311 @kyleniemeyer
@nicoguaro ok, thanks. @philipperenard we may need your help in finding an appropriate reviewer, who also has access to Matlab+necessary toolbox(es).

@pboesu, are you able to review this submission?
