### 2017-07-12 #311 @philipperenard
@kyleniemeyer. Ok. Did you receive the list of potential reviewers? I send it in response to your message 2 days ago, but I did it directly from my email and I am unsure if you got it. In this list, I suggested three names, I am pretty sure that the second one has access to matlab. Here is again the same name plus a new one and a repetition of the first.

Martin H. Trauth: http://www.geo.uni-potsdam.de/mitarbeiterdetails/show/108/Martin+H._Trauth.html - I am sure that he uses matlab. He wrote a book on Matlab for earth sciences

Alan D. Chave: http://www.whoi.edu/science/AOPE/people/achave/Site/Blank.html - He just published a book on statistics for the earth sciences using matlab.

Alex Cheng: http://home.olemiss.edu/~acheng/  - I am not sure if he has or uses matlab, but he knows very well the well testing domain. He wrote a mathematica toolbox on that topic some years ago.

Let me know if this is sufficient or if you need more help.
