### 2017-07-12 #311 @pboesu
@kyleniemeyer I am in principle happy to review this. However, two possible constraints:
1)  I'm about to do a fair amount of traveling, so realistically I will not be able to work on this review until the second half of August. If that should be a problem, I am happy to ceed to someone else.

2) My institution's licensed only for MATLAB R2014b (including the statistical toolbox). I managed to follow the installation instructions for this toolbox, and found the "Getting started" example, but am getting errors when executing parts of it (see here: https://gist.github.com/pboesu/460e5a9a046ca56ab199d838a2e6600a) . It would be great if @philipperenard could confirm whether R2017a is a hard requirement.
