### 2017-07-12 #311 @kyleniemeyer
Thanks for responding @pboesu! If we can't find someone else, then that timeline would be acceptable.

@philipperenard yes, please confirm if this is a hard requirement—and if so, one preliminary review suggestion might be to see if this requirement could be lessened, as it would enable more (potential) reuse.
