### 2017-07-13 #311 @philipperenard
Hi @kyleniemeyer, in principle, it should work with an older version. It was working with all previous versions since 2001.

@pboesu: I have seen the error message that you get and I do not have yet enough information to reproduce the problem and analyze its origin. My first guess is that the data file is not imported properly. It could be a problem of file encoding between different OS and not a problem related to the version of matlab. Which OS are you using ?

To try to reproduce the bug, I have started to download the release R2014b and will check if I obtain the same issue as you.

But, please note that I am in holiday and traveling, I'll do my best to answer rapidly. But the internet connection is slow, and takes for ever to download matlab and do the tests...
