### 2017-07-13 #311 @philipperenard
I confirm that the ths_dmo example works (as well as war_dmo, thc_dmo, grf_dmo or blt_dmo) on R2014b on Windows 10.
