### 2017-10-27 #441 @whedon
[REVIEW]: Hytool: an open source matlab toolbox for the interpretation of hydraulic tests using analytical solutions
**Submitting author:** @philipperenard (<a href="http://orcid.org/0000-0003-4504-435X">Philippe Renard</a>)
**Repository:** <a href="https://github.com/UniNE-CHYN/hytool" target ="_blank">https://github.com/UniNE-CHYN/hytool</a>
**Version:** 2.05

The repository contains a MIT license at the top level, but a GNU GPL v3 in the package documentation. Please harmonize the license terms.

The git repository consists of a single large commit of the entire software toolbox, and then some smaller commits related to the JOSS submission, so it is not possible to see how the code base grew. However, I do not see any reason to doubt that the software was created under the lead of @philipperenard. The [Contents.m](https://github.com/UniNE-CHYN/hytool/blob/93b21470e374dc57c4b63e412d5a678e4658a9d2/Contents.m#L4) file lists 'Philippe Renard and Co.' as authors, the 'and Co.' are explicitely listed in the Acknowledgments section of the toolbox documentation.

There are no automated tests as far as I can see, and as far as I can see no use is made of MATLABS [testing framework](https://www.mathworks.com/help/matlab/matlab-unit-test-framework.html)

There are no such guidelines in the MATLAB documentation or the github readme or wiki.

There are no such guidelines.

DOI is missing missing for one reference.
