### 2017-11-02 UniNE-CHYN/hytool#1 @pboesu
[JOSS review] Conflicting licenses
The repository contains a MIT license at the top level (https://github.com/pboesu/hytool/blob/master/LICENSE), but a GNU GPL v3 in the package documentation (https://github.com/pboesu/hytool/blob/master/html/license.txt) . Please harmonize the license terms.
