### 2017-11-02 UniNE-CHYN/hytool#3 @pboesu
[JOSS review] lack of unit tests
There are no automated tests as far as I can see, and as far as I can see no use is made of MATLAB's testing framework.

I am not familiar enough with MATLAB toolboxes or the MATLAB community to know whether or not this is standard practice.
