### 2017-11-02 UniNE-CHYN/hytool#4 @pboesu
[JOSS review] Lack of community guidelines
JOSS considers it good practice to provide communite guidelines for software users who want to
1) Contribute to the software
2) Report issues or problems with the software
3) Seek support

Please consider adding such guidelines to the documentation.
