### 2017-11-02 #441 @pboesu
Hello @kyleniemeyer and @philipperenard,

thank you both for your patience, and my apologies again for delaying this review.

I have now reviewed the paper and code. This is a very well documented toolbox, and all of the example cases are detailed and work.

I have highlighted a few minor issues, but these should be easy to address

1. There are conflicting licenses in the toolbox (MIT vs GPL3). Please use a single license: https://github.com/UniNE-CHYN/hytool/issues/1
4. Missing DOI in the paper. I've opened a pull request to add a DOI for one of the references cited in the paper. https://github.com/UniNE-CHYN/hytool/pull/2
2. There are no unit tests. However, I am not familiar enough with MATLAB toolboxes or the MATLAB community to know whether or not this is standard practice, so I am neutral about whether this would be a requirement for publication in JOSS, and I would like to defer to @kyleniemeyer for further guidance on this. https://github.com/UniNE-CHYN/hytool/issues/3
3. Lack of community guidelines. https://github.com/UniNE-CHYN/hytool/issues/4
