### 2017-11-05 UniNE-CHYN/hytool#1 @philipperenard
Thank you for spoting this issue. I realized that there were also some other inconsistencies in previous pieces of the code which were refering to the GNU license that I initially used (most of the dmo files). I searched systematically and replaced the obsolete references to the GNU license. Now, it should be fine.
