### 2017-11-05 UniNE-CHYN/hytool#3 @philipperenard
I was not aware of the existence of the testing framework existing in matlab. As far as I know, it is not yet a standard procedure. But, I believe that it should change. I did a first test and it seems to work very nicely. I will implement it progressively, but it will take me some time. I may not be able to make this change rapidly.
