### 2017-11-05 UniNE-CHYN/hytool#4 @philipperenard
This is a very good suggestion. I have added such guidelines in the Wiki pages. I am not very familiar of this type of information, I tried to provide the main information. Let me know if you have additional suggestions.
