### 2017-11-05 #441 @philipperenard
Thank you very much for the evaluation and suggestions. I adressed alll the points mentioned in the review and replied directly on the github links.

The only thing that I have not done yet is the to add unit tests. As I wrote in my reply, I will not be able to do them rapidly but I agree with the reviewer that they should be added.
