### 2017-11-06 #441 @pboesu
Thank you @philipperenard for the revisions!

Automated test would be great for future development, but given the very detailed worked examples, I think the review guidelines (which state that "manual steps described so that the function of the software can be verified" are sufficient) are satisfied.

@kyleniemeyer I am satisfied with the revisions and recommend the acceptance of this submission.
