### 2017-11-06 #441 @kyleniemeyer
@pboesu thanks for your review!

To both of you, JOSS does not require an automated test suite, but it is definitely encouraged in general as a best practice. However, we do want sufficient manual tests/examples to verify the functionality, and it sounds like that is satisfied.
