### 2017-11-07 #441 @kyleniemeyer
Hi @philipperenard, I'm just taking a look at the actual paper now.

I notice that you have a great figure in your wiki that demonstrates a use case for the software. Could you incorporate that, along with an explanation (similar to what is also in the wiki) into the paper? I think that would help strengthen the paper, and make it more clear how the software can be used.

Also, a few additional minor comments/suggestions:
- "allowing to identify aquifer types" -> "used to identify aquifer types..."
- Could you perhaps add a sentence after the first that explains what hydraulic tests are, to a general reader?
- "Hytool is a matlab toolbox providing" -> "Hytool is a Matlab toolbox that provides"
- "for the interpretation of hydraulic tests" -> "for interpreting hydraulic tests"
- "allows adding easily new solutions" -> "allows a user to easily add new solutions"
- "internal help within matlab" -> "internal help within Matlab"
- regarding the "several research projects", could you add some detail on how it was used? That would help people further see the use cases for their own work.
