### 2017-11-12 #441 @philipperenard
Dear @kyleniemeyer,

Thank you for your suggestions. They are very much appreciated.

Here is what I did:
1. Add the same figure as in the wiki
2. Add a few sentences in the 4th paragraph explaining the figure.
3. Corrected "allowing to identify aquifer types" -> "used to identify aquifer types..."
4. Add a sentence explaining in the first paragraph whatis a  hydraulic test
5. Corrected "Hytool is a matlab toolbox providing" -> "Hytool is a Matlab toolbox that provides"
6. Corrected "for the interpretation of hydraulic tests" -> "for interpreting hydraulic tests"
7. Corrected "allows adding easily new solutions" -> "allows a user to easily add new solutions"
8. Corrected  "internal help within matlab" -> "internal help within Matlab"
9. The last paragraph has been divided in two parts to split the text related to the help and the description of the exemples. In this last part, I added a few references and some details about "several research projects" to indicate briefly in which contect hytool was used.
