### 2017-11-12 #441 @kyleniemeyer
Thanks @philipperenard, a caught a few other things (or new one):

- "They consist in" -> "They consist of"
- "Many types of tests involving" -> "Many types of tests involve" (unless I misunderstand your meaning)

Once those are fixed, the submission will be ready to accept!
