### 2017-11-12 #441 @philipperenard
Thanks @kyleniemeyer !

I corrected the first point.

For the second, the sentence is a bit complicated, but the meaning was "Many types of tests .... had been developped...". I added some comas to clarify the structure of the sentence. In my mind, "involving...." is a parenthetic clause. I am not 100% sure that what I did  is grammaticaly correct.  I could try to split the sentence.
