### 2017-11-12 #441 @arfon
@philipperenard - please could you merge this PR: https://github.com/UniNE-CHYN/hytool/pull/5
