### 2017-11-13 #441 @philipperenard
Thank you all @pboesu , @kyleniemeyer and @arfon for all your work and advices.

What you do with this journal is fantastic ! I have been editor of a traditional journal, and the procedures and way you manage this journal should inspire all of us to move into an open framework.
